backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy
psycopg[binary]
python-dotenv==1.0.1
pydantic
